==> ipynb/3.Get_Target_Data.ipynb <==
# Part 3 -- Get Target Data (S&P 500)

Get our target data (the S&P 500 index) from the Yahoo Finance website and perform calculations. We want to look at the Adj Close price to determine whether stocks went up/down/neutral compared to the previous day. This will be a **classification model** and we will set our threshold of up/down/neutral to a change of 0.01%.

**Load lib codes**

from os import chdir
chdir('/home/jovyan/work/Analyzing_Unstructured_Data_for_Finance/')

from lib import *
# suppress_warnings()

**Convert strings to datetime and integers so we can perform calculations**

sp500 = pd.read_csv('../Analyzing_Unstructured_Data_for_Finance/data/3.^GSPC.csv')

sp500['Adj Close'][1773]

def convert_to_datetime_int(df):
    date_list = []
    for d in sp500['Date']:
        date = dt.datetime.strptime(d, '%Y-%m-%d')
        date_list.append(date)
        
    df['Date'] = pd.Series(date_list)
    df['Adj Close'] = pd.to_numeric(df['Adj Close'], errors='coerce', downcast='float')
    return df

convert_to_datetime_int(sp500)

sp500.info()

```
# Date range for tweets
# 2008-09-10 00:00:00
# 2017-06-12 00:00:00
```
print(min(sp500['Date']))
print(max(sp500['Date']))

# NOTE: Can always change this from a classification model to regression model by using Percent_Change as our target instead of Percent_Change_Class

**We want the <u>PERCENT CHANGE</u> of each stock so that data is normalized. This is especially important if you ever want to compare different stocks/indices.**

def get_change_in_Adj_Close(df):
    df = df.sort(columns='Date')
    df['Diff'] = df['Adj Close'].diff()
    df['Percent_Change'] = (df['Adj Close']-df['Adj Close'].shift(1))/df['Adj Close']
    return df

sp500_df = get_change_in_Adj_Close(sp500)

sp500_df.head(2)

# Download data and double-check results in Excel to see if calculations were executed correctly
sp500_df.to_csv('../Analyzing_Unstructured_Data_for_Finance/data/3.sp500.csv')

# Look at the distribution of Percent_Change to determine what threshold to use
print('min: ', sp500_df['Percent_Change'].min())
print('max: ', sp500_df['Percent_Change'].max())
print('mean: ', sp500_df['Percent_Change'].mean())

print('mode: ', "{0:.10f}".format(round(sp500_df['Percent_Change'],3).mode()[0]))
print('mode: ', "{0:.10f}".format(round(sp500_df['Percent_Change'],5).mode()[0]))

# Make threshold .01% (for percent change in daily Close price)
def make_binary(data):
    data_list = []
    for d in data:
        if round(d,5) < 0.001:
            data_list.append('down')
        elif round(d,5) == 0.001:
            data_list.append('neutral')
        elif round(d,5) > 0.001:
            data_list.append('up')
        else:
            data_list.append('n/a')
    return data_list

sp500_df['Percent_Change_Class'] = make_binary(sp500_df['Percent_Change'])

**We will set our threshold to .01% because a threshold of .05% (like below) gives us the following distribution, which is a little uneven:**
```
down    1484
up       526
```

sp500_df['Percent_Change_Class'].value_counts()

sp500_df.sample(2)

sp500_df.shape

joblib.dump(sp500_df, '../Analyzing_Unstructured_Data_for_Finance/data/3.sp500_df.pickle')
