==> ipynb/4.Combine_Features_And_Target.ipynb <==
# Part 4 -- Combine Features and Target

Since our feature data (tweets) and target data (stocks) are pulled from different sources and have different numbers of rows and columns, we will need to do some cleaning to make sure our X's and y's share the same number of rows. This is important for when we feed our data into models.

**Load lib codes**

from os import chdir
chdir('/home/jovyan/work/Analyzing_Unstructured_Data_for_Finance/')

from lib import *
# suppress_warnings()

sp500_df = joblib.load('../Analyzing_Unstructured_Data_for_Finance/data/3.sp500_df.pickle')

tweets_df = joblib.load('../Analyzing_Unstructured_Data_for_Finance/data/2.tweets_df.pickle')

sp500_df.shape

tweets_df.shape

**Do some checks to see if everything matches correctly before merging the dataframes**

sp500_df[2000:2002]

tweets_df.head(4)

tweets_df['Date'][3]==sp500_df['Date'][2001]

def match_dates_and_pull_y(features_df, target_df):
    change = []

    for i in features_df['Date']:
        try:
            if i in list(target_df['Date']):
                change.append(target_df['Percent_Change_Class'].loc[target_df['Date']==i].values)
            elif i not in list(target_df['Date']):
                change.append('x')
                pass
        except Exception as e:
            print('Error:', e)
            
    return pd.DataFrame(change)

start = datetime.now()

change_df = match_dates_and_pull_y(tweets_df, sp500_df)

end = datetime.now()
print(end - start)

change_df[0].value_counts()

43367+33891+17514+84

**Now we have our X's and y's**<br>
Our <u>Features</u> == Tweets (\_id, text, timestamp, username, cleaned_text, Date)<br>
Our <u>Target</u> == change in stock prices (up/down/neutral)<br>

They have the same number of rows and are in the same order. Now, we just need to drop the rows where our X's and y's don't have 'Date' in common.

**Merge & drop rows where there's no stock data for Tweets**

combined_df_nodrop = tweets_df.merge(change_df, left_index=True, right_index=True)

combined_df_nodrop.head(2)

combined_df = combined_df_nodrop[combined_df_nodrop[0]!='x']

combined_df = combined_df[combined_df[0]!='n/a']

combined_df.head(2)

joblib.dump(combined_df, '../Analyzing_Unstructured_Data_for_Finance/data/4.combined_df.pickle')

**Now split the data into X's and y's (our features and target)**

X = combined_df.drop(0, axis=1)
y = combined_df[0]

print(X.shape)
print(y.shape)

77258.0/94856.0
# We kept 81% of original tweets (after dropping ones that don't match)

joblib.dump(X, '../Analyzing_Unstructured_Data_for_Finance/data/4.X.pickle')

joblib.dump(y, '../Analyzing_Unstructured_Data_for_Finance/data/4.y.pickle')

combined_df[0].value_counts()
