==> ipynb/5.Transform_Target_Data.ipynb <==
# Part 5 -- Transform Target Data (SP500)

LabelEncode our y's so we can have more flexibility running it through models (e.g. LogisticRegression).

**Load lib codes**

from os import chdir
chdir('/home/jovyan/work/Analyzing_Unstructured_Data_for_Finance/')

from lib import *
# suppress_warnings()

y = joblib.load('../Analyzing_Unstructured_Data_for_Finance/data/4.y.pickle')

from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()

y_le = le.fit_transform(y)

le.classes_

pd.DataFrame(y_le, le.inverse_transform(y_le)).sample(5)

**Remember the following classes and their encoding**<br>
```
down == 0
up == 1
```

joblib.dump(y_le, '../Analyzing_Unstructured_Data_for_Finance/data/5.y_le.pickle')
