==> ipynb/6.2.Model_Selection_Pipelines-TESTS.ipynb <==
# Part 6 -- Model Selection

Using pipelines to find the best predictive model for our data.

**Load lib codes**

from os import chdir
chdir('/home/jovyan/work/Analyzing_Unstructured_Data_for_Finance/Analyzing_Unstructured_Data_for_Finance/')

from lib import *
# suppress_warnings()

**Try a bunch of naive models to see which works best with our data. Then, we will GridSearch over our final model and tune its hyperparameters.**

!pip install pymongo

X = joblib.load('../Analyzing_Unstructured_Data_for_Finance/data/4.X.pickle')

y = joblib.load('../Analyzing_Unstructured_Data_for_Finance/data/5.y_le.pickle')

X = X['cleaned_text']

**Before transforming any data, make sure to train_test_split so your test data doesn't bleed into your train data**

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=.2, random_state=42)

X_train.shape

y_train.shape

**Transform X's (tweets) into vectors using TF-IDF**

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf = TfidfVectorizer(ngram_range=(1,3), min_df=20, stop_words='english')
X_train_tfidf = tfidf.fit_transform(X_train)
X_test_tfidf = tfidf.transform(X_test)

joblib.dump(tfidf, '../Analyzing_Unstructured_Data_for_Finance/data/6.tfidf_transformer.pickle')

joblib.dump(X_train_tfidf, '../Analyzing_Unstructured_Data_for_Finance/data/6.X_train.pickle')

joblib.dump(y_train, '../Analyzing_Unstructured_Data_for_Finance/data/6.y_train.pickle')

joblib.dump(X_test_tfidf, '../Analyzing_Unstructured_Data_for_Finance/data/6.X_test.pickle')

joblib.dump(y_test, '../Analyzing_Unstructured_Data_for_Finance/data/6.y_test.pickle')

# X_tfidf = tfidf.transform(X)

# joblib.dump(X_tfidf, '../Analyzing_Unstructured_Data_for_Finance/data/6.X_tfidf')

from sklearn.model_selection import GridSearchCV, StratifiedShuffleSplit
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier

param_dict = {}

model_dict = {
    'mnb':GridSearchCV(MultinomialNB(),
                             param_grid=param_dict,
                             cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
    'lr':GridSearchCV(LogisticRegression(),
                             param_grid=param_dict,
                             cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
    'rf':GridSearchCV(RandomForestClassifier(),
                      param_grid=param_dict,
                      cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
    'svc':GridSearchCV(SVC(),
                      param_grid=param_dict,
                      cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
    'knn':GridSearchCV(SVC(),
                      param_grid=param_dict,
                      cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
}

def fit_all_models(X,y, model_dict):
    for model in model_dict.keys():
        model_dict[model].fit(X,y)
        print("{:5} best score: {}".format(model, model_dict[model].best_score_))

```
mnb   best score: 0.5758291538585989
knn   best score: 0.5595534703122472
rf    best score: 0.5734994337485844
lr    best score: 0.576961656689856
```

```
mnb   best score: 0.5758291538585989
knn   best score: 0.5617214043035109
rf    best score: 0.5690664940948067
lr    best score: 0.5679339912635496
svc   best score: 0.5617214043035109
1:36:53.311107
```

# start = datetime.now()

# fit_all_models(X_train, y_train, model_dict)

# end = datetime.now()
# print(end - start)

# start = datetime.now()

# fit_all_models(X_train, y_train, model_dict)

# end = datetime.now()
# print(end - start)

# RUN OVERNIGHT

param_dict = {
    'mnb':{
        'alpha': np.linspace(.1,1,5)},
    'lr':{
        'C': np.logspace(-3,3,7)},
    'rf':{
        'max_depth': [5,10,20,30,40],
        'min_samples_leaf': [1,2,5,8],
        'criterion': ['gini','entropy']},
    'svc':{
        'kernel': ['linear', 'poly', 'rbf', 'sigmoid', 'precomputed']},
    'knn':{
        'n_neighbors': np.arange(5,50,3)}
}

model_dict = {
    'mnb':GridSearchCV(MultinomialNB(),
                             param_grid=param_dict['mnb'],
                             cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
    'lr':GridSearchCV(LogisticRegression(),
                             param_grid=param_dict['lr'],
                             cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
    'rf':GridSearchCV(RandomForestClassifier(),
                      param_grid=param_dict['rf'],
                      cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
    'svc':GridSearchCV(SVC(),
                      param_grid=param_dict['svc'],
                      cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
    'knn':GridSearchCV(KNeighborsClassifier(),
                      param_grid=param_dict['knn'],
                      cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
}

start = datetime.now()

fit_all_models(X_train_tfidf, y_train, model_dict)

end = datetime.now()
print(end - start)

from sklearn.pipeline import Pipeline
from sklearn.decomposition import TruncatedSVD

tfidf_lr_pipe = Pipeline([
    ('tfidf', TfidfVectorizer(ngram_range=(1,3), min_df=20, stop_words='english')),
    ('lr', LogisticRegression(C=1E10))
])


tfidf_lr_pipe.fit(X_train, y_train)

tfidf_lr_pipe.score(X_test, y_test)

**Logistic Regression on tfidf_svd data**

tfidf_svd_lr_pipe = Pipeline([
    ('tfidf', TfidfVectorizer(ngram_range=(1,3), min_df=20, stop_words='english')),
    ('svd', TruncatedSVD(n_components = 10, random_state=42)),
    ('lr', LogisticRegression(C=1E10))
])


tfidf_svd_lr_pipe.fit(X_train, y_train)

tfidf_svd_lr_pipe.score(X_test, y_test)

tfidf_nb_pipe = Pipeline([
    ('tfidf', TfidfVectorizer(ngram_range=(1,3), min_df=20, stop_words='english')),
    ('nb', MultinomialNB())
])


tfidf_nb_pipe.fit(X_train, y_train)

tfidf_nb_pipe.score(X_test, y_test)
